`created` create contact from `updates` - contact - hero section - navbar
