+++ src/components/contact.tsx
@@ -1,72 +1,24 @@
 import { ContactData } from "@/constant/contact-data";
-import { ContactFormData } from "@/constant/contact-from-data";
-import React, { useState } from "react";
-import Input from "./input ";
-import Button from "./button/button";
-import Arrow from "@/../public/icons/Arrow.svg";
-import { ContactSchema } from "@/schema/contactSchema";
-import { useFormik } from "formik";
-import toast from "react-hot-toast";
+import React from "react";
+import ContactFrom from "./contact-from";
 
-const initialValues = {
-  firstName: "",
-  email: "",
-  lastName: "",
-  message: "",
-};
 const ContactUs = () => {
-  const [isPending, setIsPending] = useState<boolean>(false);
-
-  const formik = useFormik({
-    initialValues,
-    validationSchema: ContactSchema,
-    onSubmit: async (values) => {
-      setIsPending(true);
-      try {
-        const response = await fetch("/api/sendEmail", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify(values),
-        });
-
-        await response.json();
-
-        resetForm();
-
-        if (response.ok) toast.success("Email sent successfully!");
-        else toast.error("Failed to send email. Please try again.");
-
-        setIsPending(false);
-      } catch (error) {
-        console.error("Error:", error);
-        toast.error("An error occurred. Please try again.");
-        setIsPending(false);
-      }
-    },
-  });
-
-  const { values, errors, handleSubmit, handleChange, touched, resetForm } =
-    formik;
-
   return (
     <div
       id="contact"
-      className="flex flex-col justify-around gap-y-9 items-center p-2 rounded-2xl border border-primary-Normal"
+      className="flex flex-col justify-around gap-y-9 items-center p-2 rounded-2xl"
     >
       {/* 1st section */}
       <div className="flex flex-col justify-around items-center gap-y-3 py-6">
         <h1 className="text-primary-Normal text-5xl">Contact Us</h1>
-        <p className="text-center">
-          Have a question or need assistance? We are here to help <br />
-          you every step of the way. Get in touch with us and lets build
-          <br />
-          something amazing together.
+        <p className="text-center ">
+          Have a question or need assistance? We are here to help you every step
+          of the way. Get in touch with us and lets build something amazing
+          together.
         </p>
       </div>
       {/* 2nd section */}
-      <div className="grid sm:grid-cols-2 md:grid-cols-3 justify-items-center items-center w-full">
+      <div className="grid sm:grid-cols-2 md:grid-cols-3 justify-items-center items-center w-full gap-y-8">
         {ContactData.map((item, i) => (
           <div
             key={i}
@@ -82,48 +34,10 @@ const ContactUs = () => {
       </div>
 
       {/* 3rd section */}
-
-      <div className="grid grid-cols-1 bg-white text-black w-full max-w-3xl sm:p-5 rounded-3xl gap-y-3 p-2">
-        <div className="grid grid-cols-1 sm:grid-cols-2 gap-2">
-          {ContactFormData.map((item, i) => (
-            <div
-              key={i}
-              className={`${item.type == "email" ? `md:!col-span-2` : ``}`}
-            >
-              <Input
-                {...item}
-                value={values[item.name as keyof typeof values]}
-                error={errors[item.name as keyof typeof errors]}
-                touched={touched[item.name as keyof typeof touched]}
-                onChange={handleChange}
-              />
-            </div>
-          ))}
-        </div>
-        <div className="mb-4">
-          <label htmlFor="message" className="leading-7 text-md">
-            Send us a message
-          </label>
-          <textarea
-            value={values.message}
-            onChange={handleChange}
-            id="message"
-            name="message"
-            className="w-full bg-white-darkWhite rounded-lg border border-gray-300 focus:border-primary focus:ring-2 p-2
-             focus:ring-indigo-200 h-32 text-base outline-none text-gray-700  resize-none leading-6 transition-colors duration-200 ease-in-out"
-          ></textarea>
-        </div>
-        <div>
-          <Button
-            text="Send Message"
-            icon={Arrow}
-            onClick={handleSubmit}
-            loading={isPending}
-          />
-        </div>
-      </div>
+      <ContactFrom />
     </div>
   );
 };
 
 export default ContactUs;
+// border-primary-Normal
