+++ src/components/navbar.tsx
@@ -8,7 +8,8 @@ const Navbar = () => {
   return (
     <nav
       id="top"
-      className="text-white bg-primary rounded-t-[20px] p-5 flex flex-col sm:flex-row justify-start sm:justify-between sm:items-center gap-y-5"
+      className="text-white bg-primary rounded-t-[20px] p-5 flex flex-col sm:flex-row justify-start sm:justify-between
+       sm:items-center gap-y-5"
     >
       <Link href={"#top"} className="text-2xl font-charm">
         Sameer Ali
@@ -28,8 +29,6 @@ const Navbar = () => {
       <a href="/cv/Sameer-ali-cv.pdf" download>
         <Button text="Sameer Ali CV" icon={Arrow} />
       </a>
-
-      {/* <p>Let&#39;s Talk</p> */}
     </nav>
   );
 };
